--- src/components/DesignSkillCard/DesignSkillCard.module.css
@@ -1,41 +1,0 @@
-.skill-card-icon-design {
-    display: grid;
-    direction: rtl;
-    grid-template-columns: repeat(3, 1fr);
-    div {
-        opacity: 0.8;
-        height: 25px;
-        width: 25px;
-        &:nth-child(1) {
-            --trans: translatey(15px) skewX(-10deg);
-            --start: -2px;
-            --end: -10px;
-            transform: translatey(15px) skewX(-10deg);
-            background-color: #0F521A;
-            animation: breath 5s linear infinite alternate;
-        }
-        &:nth-child(2) {
-            --trans: translatey(10px) skewX(-10deg);
-            transform: translatey(10px) skewX(-10deg);
-            background-color: #136B22;
-
-        }
-        &:nth-child(3) {
-            --trans: translatey(5px) skewX(-10deg);
-            --start: 2px;
-            --end: 10px;
-            transform: var(--trans);
-            background-color: #3C924A;
-            animation: breath 5s linear infinite alternate;
-        }
-    }
-}
-
-@keyframes breath {
-    0% {
-        transform: var(--trans) translatex(var(--start))
-    }
-    100% {
-        transform: var(--trans) translatex(var(--end));
-    }
-}
