--- src/components/ProjectManagmentSkillCard/ProjectmanagmentSkillCard.module.css
@@ -1,21 +1,0 @@
-.skill-card-icon-pm {
-    display: grid;
-    gap: 10px;
-    grid-template-columns: repeat(3, 1fr);
-    div {
-        height: 20px;
-        width: 20px;
-        animation: glow 4s ease-in-out infinite alternate;
-    }
-}
-
-
-@keyframes glow {
-    0% {
-        box-shadow: 0 0 10px rgba(144, 238, 144, 0.5);  
-    }
-
-    100% {
-        box-shadow: 0 0 25px rgba(144, 238, 144, 0.5);
-    }
-}
