--- src/components/SoftwareEngineerSkillCard/SoftwareEngineerSkillCard.module.css
@@ -1,13 +1,0 @@
-.skill-card-icon-se {
-    animation: rotate 5s ease-in-out infinite alternate;
-}
-
-
-@keyframes rotate {
-    0% {
-        transform: rotate(-50deg)
-    }
-    100% {
-        transform: rotate(-20deg);
-    }
-}
