+++ src/components/SoftwareEngineeringSkillCard/SoftwareEngineerSkillCard.tsx
@@ -1,12 +1,10 @@
-import { cn } from "@/utils/cn";
 import {
   SkillCard,
   SkillCardDescription,
   SkillCardHeader,
-  SkillCardIcon,
+  SkillCardIconSE,
   SkillCardTitle,
 } from "../SkillCard/SkillCard";
-import styles from "./SoftwareEngineerSkillCard.module.css";
 
 const markdown = `
 Understand software engineering best practices and maintaining a coding standard 
@@ -22,13 +20,5 @@ export const SoftwareEngineerSkillCard = () => {
       </SkillCardHeader>
       <SkillCardDescription markdown={markdown} />
     </SkillCard>
-  );
-};
-
-export const SkillCardIconSE = () => {
-  return (
-    <div className={cn(styles["skill-card-icon-se"])}>
-      <SkillCardIcon />
-    </div>
   );
 };
